fix: refactors error controller and error entities
This commit does expose controller level details into the error handling
of the domain layer. However, I think this is justfied as it vastly simplifies
the code using polymorphism. Not only that, but the platform of sym will continue
to be a future so adding http status codes to the error does not seem terrible.